refactor(api): extract task stream + events handlers
Finishes the handler_tasks.go split started in earlier PRs. The
SSE stream handler, the event-append handler, the snapshot
builder, and the shared task/run loaders + SSE plumbing move to a
sibling so each handler file owns one concern.

  handler_tasks.go:        1613 → 1147 LOC
  handler_tasks_stream.go: new, 494 LOC

What moves to handler_tasks_stream.go

  Stream + events (3 handlers)
  - HandleTaskRunStream
  - HandleTaskRunEvents
  - HandleAppendTaskRunEvent

  Snapshot/event decoders + helpers
  - buildTaskRunStreamState
  - decodeTaskRunEventData
  - decodeTurnCostFromEventData
  - parseAfterSequence

  Shared loaders + SSE plumbing
  - loadAuthorizedTask
  - loadAuthorizedTaskRun
  - writeSSEHeaders

The loaders are used by other handler files too (lifecycle, patches);
they're file-scoped to handler_tasks_stream.go but visible to the
whole package since everything is `package api`. Bundling them with
the stream handler clusters the "read a task/run with auth checks"
logic next to the file that uses them most.

What stays in handler_tasks.go

  CRUD (HandleCreateTask, HandleTasks, HandleTask, HandleDeleteTask,
  HandleStartTask), runs/steps (HandleTaskRuns, HandleTaskRun,
  HandleTaskRunSteps, HandleTaskRunStep), and the buildTaskItem /
  buildTaskActivityItems / sortTaskActivityItems / cloneActivitySummary
  helpers used by both CRUD and stream paths.

The receiver type stays *Handler everywhere; no public API change;
mux registrations in server.go are unchanged. Imports per file
pruned via goimports.

Verification

  go build ./...                — clean
  go vet ./...                  — clean
  gofmt -s -l .                 — clean
  go test ./... -race -count=1  — 1779 / 1779 pass
